workshop/modules/05/langgraph/requirements.txt
bedrock-agentcore>=1.0.5
bedrock-agentcore-starter-toolkit>=0.2.0
langchain>=1.2.0
langchain-core>=1.2.0
langchain-aws>=0.2.0
langchain-mcp-adapters>=0.2.1
langgraph>=1.0.0
langgraph-checkpoint-aws>=1.0.2
mcp>=1.16.0
pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
boto3>=1.40.67
python-dotenv>=1.0.0
opentelemetry-instrumentation-langchain>=0.42.0
opentelemetry-semantic-conventions-ai>=0.4.11
aws-opentelemetry-distro